projects/week5-realtime-crypto-pipeline/cloud_function/requirements.txt
requests==2.31.0
google-cloud-pubsub==2.21.1
google-cloud-bigquery==3.17.2